requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
mysql-connector-python==8.1.0
numpy==1.24.3
matplotlib==3.7.2
networkx==3.1
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0